Remove DOCKER_STATE env var
DOCKER_STATE env var is supposed to control the rake tasks during a
deployment, however currently it is baked into DOCKER_STATE. Presumably
this first happened for running the container locally (on dev laptop),
as with a real environment that env var is always set to none or whatever
the deploy task was.

If the developer really needs this locally then it should be incorporated
as part of an env file or a docker runtime env var instead, as it is
dangerous to have this baked in (eg if someone runs this in production
without passing an --env-file).