Document landing page redesign
Increased the space between the autosuggest dropdown and the ‘Edit document’ button.

Set document type list in alpha order.

Set document activerecord lookup to return all documents with a publication date from ‘today’ onwards.

Document landing page to display a table of documents that have a publication date set for ‘today’ onwards.

Documents listed in the ‘To be published’ table have a link to the edit function encapsulating each title.